Move the matchmakingform to 'component' directory
